requirements.txt
flask>=2.0.0
langchain>=0.1.0
langchain-pinecone>=0.2.0
pinecone-client>=2.2.0
beautifulsoup4>=4.9.0
requests>=2.25.0
python-dotenv>=0.19.0
langchain-community>=0.0.10
langchain-core>=0.1.10